+++ src/containers/scene-manager/SceneManager.jsx
@@ -1,17 +1,6 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-
 import { WebGLRenderer, PerspectiveCamera, Scene } from 'three';
-import Terrain from '../../components/scene-objects/Terrain';
-
-/* 
-    Responsible for:
-        - Creating Renderer, Scene and Camera
-        - Initialize SceneSubjects ( Meshes, Models, Lighting )
-        - Updating everything each frame
-
-    This container shouldn't know about the DOM.
-*/
+import { Terrain } from '../../components/scene-subjects/Terrain';
 
 export default class SceneManager
 extends Component {
@@ -20,49 +9,46 @@ extends Component {
         super(props);
         this.state = {
             aspect: this.props.width / this.props.height
-        };
+        }
 
-        this.updateSceneHandler = this.updateSceneHandler.bind(this);
-        this.renderSceneHandler = this.renderSceneHandler.bind(this);
+        this.componentStartRender = this.componentStartRender.bind(this)
+        this.componentAddSubjects = this.componentAddSubjects.bind(this)
     }
 
-    componentWillMount() {
+    componentDidMount() {
         this.scene = new Scene();
         this.camera  = new PerspectiveCamera( 75, this.state.aspect, 1, 1000);
         this.renderer = new WebGLRenderer();
-        this.objects = createSceneObject(this.scene);
+
+        this.componentAddSubjects();
+
+        this.mount.appendChild(this.renderer.domElement);
     }
 
-    updateSceneHandler() {
-        for ( let i = 0; i < this.sceneObject.length; i++) {
-            sceneObject[i].update();
-        }
-
-        this.renderer.render( this.scene, this.camera);
-    }
-
-    renderSceneHandler() {
-        this.renderer.render(this.scene, this.camera);
-        requestAnimationFrame(this.renderSceneHandler);
-    }
-
-    createSceneObject(scene) {
-        const sceneObject = [
-            new Terrain(scene)
+    componentAddSubjects() {
+        this.sceneSubjects = [
+            new Terrain(this.scene)
         ];
 
-        return sceneObject;
+        return this.sceneSubjects;
+    }
+
+    componentLoadSubjects() {
+        this.sceneSubjects = componentAddSubjects();
+        for ( let i = 0; i < this.sceneSubjects.length; i++ ) {
+            this.sceneSubjects[i].update();
+        }
+
+        this.renderer.render(this.scene, this.camera);
+    }
+
+    componentStartRender() {
+        this.renderer.render( this.scene, this.camera );
+        requestAnimationFrame( this.componentStartRender );
     }
 
     render() {
-        requestAnimationFrame(this.render);
-        updateSceneHandler();
-        return <div ref={ (mount) => { this.mount = mount } }></div>
+        return <div style={{ width: '400px', height: '400px' }} ref={ (mount) => { this.mount = mount } }></div>
     }
 
-}
-
-SceneManager.propTypes = {
-    height: PropTypes.number.isRequired,
-    width: PropTypes.number.isRequired
 }
